Add .NET Standard support to ColorMine
Move color profile support (which requires WPF and PInvoke calls) to a separate assembly.
Write HSL and HSV converters from scratch (to remove dependency on WPF / System.Drawing)
