+++ ColorMine.Test/ColorMine.Test.csproj
@@ -1,6 +1,6 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net45</TargetFrameworks>
+    <TargetFrameworks>netcoreapp1.1</TargetFrameworks>
     <RootNamespace>ColorMine.Test</RootNamespace>
     <TestProjectType>UnitTest</TestProjectType>
   </PropertyGroup>
