+++ ColorMine.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ColorMine", "ColorMine\Colo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ColorMine.Test", "ColorMine.Test\ColorMine.Test.csproj", "{318DFB24-B947-4B07-AFDB-561A262EB424}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ColorMineProfiles", "ColorMineProfiles\ColorMineProfiles.csproj", "{24F842DC-B627-45C8-8CFD-CFABBEA34F5A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{318DFB24-B947-4B07-AFDB-561A262EB424}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{318DFB24-B947-4B07-AFDB-561A262EB424}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{318DFB24-B947-4B07-AFDB-561A262EB424}.Release|Any CPU.Build.0 = Release|Any CPU
+		{24F842DC-B627-45C8-8CFD-CFABBEA34F5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{24F842DC-B627-45C8-8CFD-CFABBEA34F5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{24F842DC-B627-45C8-8CFD-CFABBEA34F5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{24F842DC-B627-45C8-8CFD-CFABBEA34F5A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
