+++ ColorMineProfiles/ColorMineProfiles.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net45</TargetFrameworks>
+    <AssemblyTitle>ColorMine.Profiles</AssemblyTitle>
+    <RootNamespace>ColorMine</RootNamespace>
+    <Version>1.2.0</Version>
+    <FileVersion>1.2.0</FileVersion>
+    <Description>Adds ICC profile support for ColorMine</Description>
+    <Company>Continuous Automation LLC</Company>
+    <Copyright>Copyright © 2017 Joe Zack</Copyright>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ColorMine\ColorMine.csproj" />
+    <Reference Include="PresentationCore" />
+  </ItemGroup>
+
+</Project>
